W174b: THE CAMP SYSTEMIC DEFECT IS SOLVED, plus the mdlFile VU stub family completes.
_h_camp_subb 001406D0 (campDrawEquipmentDetailList) and 00140E30 (campDrawEquipmentDetailListAlt) both went nd1253, 1620/1888 -> **MATCH at 1880/1888**. These twins were the file's only near-solved functions and the seed of its recipe; the file was 2/22 matching, the worst rate in the project. The fix recovered their real CampPair/f32/typed-work/s32 ABI and the full sprite/text/value call shapes: u8 alpha, signed detail values, absolute-array versus GP-relative global addressing, branch-local variants, and the alternate twin's 0x1d/0x1c frames. Both callers migrated, readable aliases added. Phantom census over the slice is ZERO, confirming the family's fptodp calls came from untyped CALLEE DECLARATIONS promoting float arguments under K&R rules, not from literals.

mdlFile: the five-function VU stub family is DONE -- 00349AF0, 0034A370, 0034A9C0, 0034AE30 all reached MATCH (tags stripped after STALE confirmation) alongside 00349FD0, completing the family converted from fake placeholder macros to real VU macro-mode asm. Plus 0034D990, 00350230, 00351D20, 00355E70, 00356BA0 MATCHed and 0034D150 went 656/592 nd471 -> 588/592 nd9 with the retail VU0 sequence recovered as ONE contiguous hardware-asm block, every revision compile-verified. mdlFile is now **440 MATCH / 227 NONMATCHING (66.0%)**, up from 423.

fclShopMisc THREE MATCHes: 003F0830 (fclShopSelectStockItemId), 003F1BA0 (fclShopPopulateEquipmentEntry), 003F9E30 (fclShopCreateEquipmentItemList); 003F2240 is at nd2.

ADJUDICATION: reverted fclShopMisc 003F7C60 by marker-bounded byte splice -- it had gone 232/240 nd74 -> 280/240 nd227, crossing its window. Restored exactly. Zero lost, zero new over-window, zero COMPILE_ERROR, zero stale remaining.

+++ src/datCalendar.c
@@ -1462,7 +1462,7 @@ u8 clndIsHolidayOrSunday()
     s32 month;
     const Holiday* holidays;
     s32 day;
-    s16 sentinel;
+    s32 sentinel;
 
     daysSinceApr5 = datGetDaysSinceApr5();
     if ((daysSinceApr5 + CALENDAR_DAY_MAX) % CALENDAR_DAY_MAX == CALENDAR_DAY_SUNDAY)
